+++ contracts/modules/delegates/IexecPocoBoostDelegate.sol
@@ -71,6 +71,8 @@ contract IexecPocoBoostDelegate is IexecPocoBoost, IexecAccessorsBoost, Delegate
             _requestorder.datasetmaxprice >= _datasetorder.datasetprice,
             "PocoBoost: Overpriced dataset"
         );
+        // An intermediate variable stored in the stack consumes
+        // less gas than accessing calldata each time.
         uint256 workerpoolPrice = _workerpoolorder.workerpoolprice;
         require(
             _requestorder.workerpoolmaxprice >= workerpoolPrice,
@@ -258,6 +260,10 @@ contract IexecPocoBoostDelegate is IexecPocoBoost, IexecAccessorsBoost, Delegate
         {
             uint256 taskPrice = _apporder.appprice + _datasetorder.datasetprice + workerpoolPrice;
             lock(deal.requester, taskPrice * volume);
+            // Order is important here. First get percentage by task then
+            // multiply by volume.
+            uint256 taskWorkerpoolStake = (workerpoolPrice * WORKERPOOL_STAKE_RATIO) / 100;
+            lock(deal.workerpoolOwner, taskWorkerpoolStake * volume);
         }
         // Notify workerpool.
         emit SchedulerNoticeBoost(